backend/foodgram/requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
django-environ
et-xmlfile==1.0.1
flake8==3.9.1
idna==2.9
importlib-metadata==1.6.0
isort==5.8.0
MarkupPy==1.14
mccabe==0.6.1
more-itertools==8.2.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
PyYAML==5.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==3.0.0
urllib3==1.25.9
wcwidth==0.1.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv
djoser